# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/product/terms:import": {
            "post": {
                "tags": [
                    "Product Terms File"
                ],
                "summary": "Upload Csv File",
                "description": "Upload Csv File",
                "operationId": "Upload Csv File",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "csv_file": {
                                        "description": "Upload Csv file",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
        "/api/v1/product/terms:delete": {
            "post": {
                "tags": [
                    "Product Terms File"
                ],
                "summary": "Upload Csv File to Delete Items",
                "description": "Upload Csv File to Delete Items",
                "operationId": "Upload Csv File",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "properties": {
                                    "csv_file": {
                                        "description": "Upload Csv file",
                                        "type": "file",
                                        "format": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
        "/api/v1/product/terms:export": {
            "post": {
                "tags": [
                    "Product Terms File"
                ],
                "summary": "Export Csv File",
                "description": "Export Csv File",
                "operationId": "Export Csv File",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "export_all": {
                                        "description": "Export All",
                                        "type": "bool"
                                    },
                                    "selected_id": {
                                        "description": "Selected IDs array",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
        "/api/v1/product/terms:download/{id}": {
            "get": {
                "tags": [
                    "Product Terms File"
                ],
                "summary": "Download Exported File",
                "description": "Download Exported File",
                "operationId": "Download Exported File",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "export job Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "basic": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "basic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}